Add IItemService, IItemService
Service layer that can communicate with the DB context class and
fetch/write data to the/from the database. We used
an interface-implementation model for this.
Create a new Folder Services and add in a new interface IItemService.
This basically defines what our Item Service is supposed to perform.(CRUD)